Support parsing raw EXIF data
This adds support for autodetection and parsing of raw exif data.
If the data passed to Exif::Data.new begins with the EXIF Header
"Exif\x00\x00" it is treated as raw EXIF data, otherwise the libexif
loader is used.

Parsing of raw EXIF data is only supported for strings.